+++ fetch3/initial_conditions.py
@@ -2,16 +2,29 @@
 ##################
 Initial conditions
 ##################
+
+Assume hydrostatic conditions in plant
+- potential at bottom equals the potential specified in the soil layer where the root bottom is
+- potential at height z = initial head at bottom or roots + rho*g*h
+
+For soil layers
+- specify an initial soil moisture for each soil layer
+- try just using a constant potential over each layer (but check to see if this causes issues in the solver).
+- might need to do some interpolation between layers
 """
 import numpy as np
 
-#######################################################################
-# INITIAL CONDITIONS
-#######################################################################
-# soil initial conditions as described in the paper [VERMA et al., 2014]
-
 
 def initial_conditions(cfg, q_rain, zind):
+
+    # clay layer
+
+    # sand layer
+
+    # roots
+    # P0 = head at the soil layer where the root bottom is [Pa]
+    # xylem
+
     dz = cfg.dz
 
     initial_H = np.zeros(shape=zind.nz)
@@ -50,20 +63,19 @@ def initial_conditions(cfg, q_rain, zind):
     ######################################################################
     soil_bottom = np.zeros(shape=len(q_rain))
     for i in np.arange(0, len(q_rain), 1):
-        soil_bottom[i] = 28  # 0.28 m3/m3 fixed moisture according to VERMA ET AL., 2014
+        soil_bottom[i] = cfg.soil_moisture_bottom_boundary
 
     # clay - van genuchten
     Head_bottom = (
         (
-            ((cfg.theta_R1 - cfg.theta_S1) / (cfg.theta_R1 - (soil_bottom / 100))) ** (1 / cfg.m_1)
-            - 1
+            ((cfg.theta_R1 - cfg.theta_S1) / (cfg.theta_R1 - soil_bottom)) ** (1 / cfg.m_1) - 1
         )
         ** (1 / cfg.n_1)
     ) / cfg.alpha_1
     Head_bottom_H = -Head_bottom * cfg.g * cfg.Rho  # Pa
-    Head_bottom_H = np.flipud(
-        Head_bottom_H
-    )  # model starts the simulation at the BOTTOM of the soil
+
+    # model starts the simulation at the BOTTOM of the soil
+    Head_bottom_H = np.flipud(Head_bottom_H)
 
     ############## inital condition #######################
     # setting profile for initial condition